+++ bioexplorer/core/plugin/fields/Octree.h
@@ -45,7 +45,7 @@ public:
      * @param minAABB Lower bound of the scene bounding box
      * @param maxAABB Upper bound of the scene bounding box
      */
-    Octree(const floats &events, float voxelSize, const glm::vec3 &minAABB,
+    Octree(const floats &events, double voxelSize, const glm::vec3 &minAABB,
            const glm::vec3 &maxAABB);
 
     /**
